Update main.py
Excellent work! I like the extensions you've done with logging in (and Bootstrap), so I'm including one modification that you'll be able to appreciate.

Because your username and password were here in your code, I was able to use it to make a donation. But instead of storing them in the code, you might prefer to hide them as configuration variables which you can set in Heroku and retrieve with the included change. To set them in heroku, you would:

```bash
$ heroku config:set USERNAME=admin
$ heroku config:set PASSWORD=admin
```